Backend/requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2==2.9.10
sqlparse==0.5.3
